backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and caching
redis==5.0.1
celery==5.3.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
pdfplumber==0.10.3

# MinIO/S3
minio==7.2.0
boto3==1.29.7

# AI/LLM
google-generativeai==0.3.2
langchain==0.0.340
langchain-google-genai==0.0.6
tiktoken==0.5.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
pydantic-settings==2.1.0